+++ lua/configuration/keybindList.lua
@@ -1,17 +1,4 @@
-local map = vim.api.nvim_set_keymap
-
--- simpler cover for vim.api.nvim_set_keymap
-local function simpleMap(modemap, keybind, command, optionsArg)
-    --[[
-        Default args include: modmap = "n", optionsArg = {noremap = true}
-    --]]
-    if not keybind then return end
-    if not command then return end
-    map(modemap or "n", keybind, "<cmd>" .. command .. "<CR>", optionsArg or {noremap = true})
-end
-
--- keybindingsTable, optional args to simpleMap function: modemap and optionsArg
-local keybindingsTable = {
+return {
     {keybind = "<leader>ff", command = "lua require('telescope.builtin').find_files()"},
     {keybind = "<leader>fg", command = "lua require('telescope.builtin').live_grep()"},
     {keybind = "<leader>fb", command = "lua require('telescope.builtin').buffers()"},
@@ -24,8 +11,3 @@ local keybindingsTable = {
     {keybind = "<leader>a", command = "lua require('persistence').load()"},
     {keybind = "<leader>v", command = "lua require('persistence').load({last=true})"},
 }
-
--- loop through the keybindingsTable and map the keys
-for _, keybindElement in ipairs(keybindingsTable) do
-    simpleMap(keybindElement.modemap, keybindElement.keybind, keybindElement.command)
-end
